pyproject.toml
[build-system]
requires = ['setuptools', 'wheel']
build-backend = "setuptools.build_meta"

[project]
name='pyimage2pdf'
dynamic = ["version"]
description = 'An adapter'
readme = "README.md"
license = {text = 'GNU AFFERO GENERAL PUBLIC LICENSE'}
authors = [{name = 'Humberto A. Sanchez II', email = 'Humberto.A.Sanchez.II@gmail.com'}]
maintainers = [{name = 'Humberto A. Sanchez II', email = 'Humberto.A.Sanchez.II@gmail.com'}]
keywords = ['pyut,images,adapter,pdf']
classifiers = [
  "Programming Language :: Python",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU Affero General Public License v3",
  "Operating System :: MacOS :: MacOS X",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  'codeallybasic>=1.8.5',
  'click>= 8.1.8',
  'pillow==11.0.0',
  'pypdf>=5.1.0',
]

[project.urls]
Repository = 'https://github.com/hasii2011/pyimage2pdf'


[tool.setuptools.packages.find]
where = ['src']

[tool.setuptools.package-data]
'pyimage2pdf.resources' = ['loggingConfiguration.json']

[tool.setuptools.dynamic]
version = {attr = 'pyimage2pdf.__version__'}

[project.scripts]

image2pdf = "pyimage2pdf.CommandClass:commandHandler"
